+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Add retry to api request
 - Search for deploy job in all pipelines
+- Add ending hot key for sync command
 
 ### Changed
 - Use cross platform notifier